20.1.Uploading References{The Upload Endpoint}
Unlike the main form on this page, this form will submit to a different endpoint. And instead of continuing to put more things into ArticleAdminController, let's create a new controller for everything related to article references: ArticleReferenceAdminController. Extend BaseController - that's just a small base controller we created in our Symfony series: it extends the normal AbstractController. So nothing magic happening there.

The Upload Endpoint

Back in the new class, create public function uploadArticleReference() and, above, @route: make sure to get the one from Symfony/Component. Set the URL to, how about, /admin/article/{id}/references - where the {id} is the Article id that we want to attach the reference to. Add name="admin_article_add_reference". Oh, and let's also set methods={"POST"}.

That's optional, but it'll let us create another endpoint later with the same URL that can be used to fetch all the references for a single article.

Let's keep going! Because the article {id} is in the URL, add an Article $article argument. Oh, and we need security! You can only upload a file if you have access to edit this article. In our app, we check that with this @IsGranted("MANAGE", subject="article") annotation, which leverages a custom voter that we created in our Symfony series. It basically makes sure that you are the author of this article or a super admin.

Finally, we're ready to fetch the file: add the Request argument - the one from HttpFoundation - and let's dd($request->files->get()) and then the name from the input field: reference.

Solid start. Copy the route name and head back to the template. Set the action attribute to {{ path() }}, the route name, and for the placeholder part, I'll use multiple lines and pass id set to article.id. Oh wait... we don't have an article variable inside this template. We do have the articleForm variable, and we could get the Article from that... but to shorten things, let's properly pass it in.

Find the edit() action of ArticleAdminController and pass an article variable. Now we can say article.id.

Phew! Ok, let's check this out: refresh and inspect element on the form. Yep, the URL looks right and the enctype attribute is there. Ok, try it: select the Symfony Best Practices doc and... upload! Yes! It's our favorite UploadedFile object!