20.2.Uploading References{Setting up UploaderHelper}
These article references are special because we need to keep them private: they should only be accessible to the author or a super admin. The process for uploading & downloading private files is, a bit different.

Setting up UploaderHelper

But, we'll start in very similar way: by opening our favorite service, and all-around nice class, UploaderHelper. Down here, add a new public function uploadArticleReference() that will have a File argument and return a string... pretty much the same as the other method, except that we won't need an $existingFilename because we won't let ArticleReference objects be updated. If you want to upload a modified file - cool! Delete the old ArticleReference and upload a new one. You'll see what I mean as we keep building this out.

To get started, just dd().

Back in the controller, let's finish this whole darn thing. Set the file to an  object and I'll add the same inline documentation that says that this is an UploadedFile object - the one from HttpFoundation.

Then say $filename =... oh - we don't have the UploaderHelper service yet! Add that argument: UploaderHelper . Then $filename = $uploaderHelper->uploadArticleReference($uploadedFile).

We know that won't work yet... but if we use our imagination, we know that... someday, it should return the new filename that was stored on the filesystem. To put this value into the database, we need to create a new ArticleReference object and persist it.