20.3.Uploading References{Tightening Up ArticleReference}
Tightening Up ArticleReference

Oh, but before create a new ArticleReference and persist it - go open that class. This is a very traditional entity: it has some properties and everything has a getter and a setter. That's great, but because every ArticleReference needs to have its Article property set... and because an ArticleReference will never change articles, find the setArticle() method and... obliterate it!

Instead, add a public function __construct() with a required Article argument. Set that onto the article property. This is an optional step - but it's always nice to think critically about your entities: what methods do you not need?

Saving ArticleReference & the Original Filename

Back up in our controller, say $articleReference = new ArticleReference() and pass $article. Call $articleReference->setFilename($filename) to store the unique filename where this file was stored on the filesystem.

But remember! There are a couple of new pieces of info that we can set on ArticleReference - like the original filename. Set that to $uploadedFile->getClientOriginalName(). Now, technically this method can return null, though, I'm not actually sure if that's something that can happen in any realistic scenario. But, just in case, add ?? . So, if the client original name is missing for some reason, fall back to $filename.

Finally, just in case we ever want to know what type of file this is, we'll store the file's mime type. Set this to $uploadedFile->getMimeType(). This can also return null - so default it to application/octet-stream, which is sort of a common way to say "I have no idea what this file is".

With that done, save this: add the EntityManagerInterface  argument, then $entityManager->persist($articleReference) and $entityManager->flush().

Finish with return redirectToRoute() and send the user back to the edit page: admin_article_edit passing this id set to $article->getId().

Yep - that's the route on the edit endpoint.

Alright! With any luck, it should hit our dd() statement. Go back to your browser: I already have the Symfony Best Practices PDF selected. Hit update... yea! UploadedFile coming from UploaderHelper.

Next: let's move the uploaded file... except that... we can't move it using the filesystem service object we have now... because we can't store these private files in the public/ directory.