22.4.Mime Type Validation{Requiring the File}
Requiring the File

Oh! There was one last case we need to validate for. Hit enter on the URL to refresh the form. If we did nothing and hit upload everything exploded inside UploaderHelper... because there was no uploaded file! The horror!

Back in the controller, the second argument to validate() can accept an array of validation constraints. Put the new File into an array. Then add: new NotBlank() with a custom message: please select a file to upload.

Refresh one more time. The huge error is replaced by a much more pleasant validation message.

Next: the author can upload a file reference... but it is literally impossible for them to download it. How can we make these private files accessible, but still check security first?