24.3.Streaming the File Download{Content-Disposition: Forcing Download}
Content-Disposition: Forcing Download

Another thing you might want to do is force the browser to download the file. It's really up to you. By default, based on the Content-Type, the browser may try to open the file - like it is here - or have the user download it. To force the browser to always download the file, we can leverage a header called Content-Disposition.

This header has a very specific format, so Symfony comes with a helper to create it. Say \ = HeaderUtils::makeDisposition(). For the first argument, we'll tell it whether we want the user to download the file, or open it in the browser by passing HeaderUtils::DISPOSITION_ATTACHMENT or DISPOSITION_INLINE.

Next, pass it the filename.

This is especially cool because, without this, the browser would probably try to call the file... just... "download" - because that's the last part of the URL. Now it will use $reference->getOriginalFilename().

Before we set this header, I just want you to see what it looks like. So, dd($disposition) move over, refresh and... there it is. It's just a string, like any other header - but it has this specific format, which is why Symfony has a helper method.