25.4.Dropzone: AJAX Upload{Customizing Dropzone}
Customizing Dropzone

Dropzone uploads a field called file. But in the controller, we're expecting it to be called reference.
We could fix this in the controller... but we can also configure Dropzone to use the reference key. We're going to do that because, in general, as cool as it is that we can just add a "dropzone" class to our form and it mostly works, to really get this system working, we're going to need to customize a bunch of things on Dropzone.

Open up admin_article_form.js. First, at the very top, add Dropzone.autoDiscover = false. That tells Dropzone to not automatically configure itself on any form that has the dropzone class: we're going to do it manually.

Try it out - close the extra tab and refresh. Hmm... still there? Maybe a force refresh? Now it's gone. The dropzone class still gives us some styling, but it's not functional anymore.