25.5.Dropzone: AJAX Upload{Customizing Dropzone}
To get Dropzone working again, inside the document.ready(), call a new initializeDropzone() function.

Copy that name, and, below, add it: function initializeDropzone(). If I were using Webpack Encore, I'd probably organize this function into its own file and import it. The goal here is to find the form element and initialize Dropzone on it. To do that, let's add another class on the form: js-reference-dropzone. Copy that, and back inside our JavaScript, say var formElement = document.querySelector() with .js-reference-dropzone.

Yes, yes, I'm using straight JavaScript here instead of jQuery to be a bit more hipster - no big reason for that. There's also a jQuery plugin for Dropzone. Next, to avoid an error on the "new" form that doesn't have this element, if !formElement, return.

Finally, initialize things with var dropzone = new Dropzone(formElement). And now we can pass an array of options. The one we need now is paramName. Set it to reference.

That should do it! Head over and select another file - how about earth.jpeg. And... cool! It looks like it worked. Click to open the profiler for the AJAX request.

Oh... careful - once again, we got redirected! So this is the profiler for the edit page. Click the link to go back to the profiler for the POST request and go back to the Debug tab. Yes! Now we're getting the normal UploadedFile object.

Close this and refresh. Look at the list! There is earth.jpeg! It worked! Of course, it's a little weird that it redirected after success... and if there were a validation error... that would also cause a redirect... and so it would look successful to Dropzone. The problem is that our endpoint isn't set up to be an API endpoint. Let's fix that next and make Dropzone read our validation errors.